# specs/openapi.yaml
openapi: 3.0.3
info:
  title: eyeO Platform API
  description: |
    Privacy-preserving video surveillance platform with Zero-Trust architecture.
    
    ## Architecture Overview
    The eyeO platform implements a Shared-Nothing Architecture (SNA) with split-brain data flows:
    - **Red Flow**: Protected video data (client-side processing only)
    - **Blue Flow**: Detection events and metadata (server-side analytics)
    
    ## Security Model
    - JWT-based authentication (HS512)
    - Client-side data transformation
    - Role-based access control (FREE/PRO/ENTERPRISE)
    - Rate limiting: 10 req/sec per user
    
    ## Base URL
    Production: `https://api.eyeo-platform.com`
    Development: `https://localhost`
    
  version: 1.0.0
  contact:
    name: eyeO Platform Security Team
    email: security@eyeo-platform.com
  license:
    name: Proprietary
    url: https://eyeo-platform.com/license
servers:
  - url: https://api.eyeo-platform.com/v1
    description: Production API
  - url: https://staging-api.eyeo-platform.com/v1
    description: Staging API
  - url: https://localhost/v1
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/api/auth/login` endpoint.
        Include in Authorization header: `Bearer <token>`
    
  # ==================== Reusable Schemas ====================
  schemas:
...
    DetectionEvent:
      type: object
      required:
        - eventType
        - cameraId
        - timestamp
      properties:
        eventType:
          type: string
          enum: [PERSON_DETECTED, VEHICLE_DETECTED, MOTION_DETECTED, ALERT]
          example: PERSON_DETECTED
        cameraId:
          type: string
          example: edge-node-001
        timestamp:
          type: integer
          format: int64
          example: 1704235200000
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.87
        boundingBox:
          $ref: '#/components/schemas/BoundingBox'
        metadata:
          type: object
          additionalProperties: true
    BoundingBox:
      type: object
      properties:
        x:
          type: integer
          example: 120
        y:
          type: integer
          example: 45
        width:
          type: integer
          example: 200
        height:
          type: integer
          example: 350
    EventBatch:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/DetectionEvent'
          maxItems: 100
    
    # Error Schemas
...
paths:
...
  /api/v1/events:
    post:
      tags:
        - Events
      summary: Ingest detection events
      description: Receives AI detection events for server-side analytics (Blue Flow)
      operationId: ingestEvents
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBatch'
      responses:
        '202':
          description: Events accepted for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: integer
                    example: 42
                  batchId:
                    type: string
                    example: batch-2024-001
        '400':
          description: Invalid event format
    
    get:
      tags:
        - Events
      summary: Query detection events
      description: Retrieve historical events with filtering
      operationId: queryEvents
      security:
        - bearerAuth: []
      parameters:
        - name: cameraId
          in: query
          schema:
            type: string
        - name: eventType
          in: query
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: integer
            format: int64
        - name: endTime
          in: query
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: Events retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/DetectionEvent'
                  total:
                    type: integer
                  page:
                    type: integer
  
  # ==================== Health Endpoints ====================
...
security:
  - bearerAuth: []